+++ src/transformers/models/data2vec_vision/convert_data2vec_vision_unilm_to_pytorch.py
@@ -1,72 +1,55 @@
-# coding=utf-8
-# Copyright 2021 The HuggingFace Inc. team.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""Convert Data2VecVision checkpoints from the unilm repository."""
-
-
+#!/usr/bin/env python3
+import torch
 import argparse
 import json
-from pathlib import Path
+from timm.models import create_model
+from transformers import (
+    BeitFeatureExtractor,
+    Data2VecVisionConfig,
+    Data2VecVisionModel,
+    Data2VecVisionForImageClassification,
+)
+from huggingface_hub import hf_hub_download
 
-import torch
-from datasets import load_dataset
 from PIL import Image
 
-import requests
-from huggingface_hub import hf_hub_download
-from transformers import (
-    Data2VecVisionConfig,
-    BeitFeatureExtractor,
-    Data2VecVisionForImageClassification,
-    Data2VecVisionForMaskedImageModeling,
-    Data2VecVisionForSemanticSegmentation,
-)
-from transformers.utils import logging
 
-
-logging.set_verbosity_info()
-logger = logging.get_logger(__name__)
-
-
-# here we list all keys to be renamed (original name on the left, our name on the right)
-def create_rename_keys(config, has_lm_head=False, is_semantic=False):
+# Copied from transformers.models.beit.modeling_beit.create_rename_keys
+def create_rename_keys(config, has_lm_head=False, is_semantic=False, hf_prefix="data2vec."):
     prefix = "backbone." if is_semantic else ""
 
     rename_keys = []
     for i in range(config.num_hidden_layers):
         # encoder layers: output projection, 2 feedforward neural networks and 2 layernorms
-        rename_keys.append((f"{prefix}blocks.{i}.norm1.weight", f"data2vec_vision.encoder.layer.{i}.layernorm_before.weight"))
-        rename_keys.append((f"{prefix}blocks.{i}.norm1.bias", f"data2vec_vision.encoder.layer.{i}.layernorm_before.bias"))
         rename_keys.append(
-            (f"{prefix}blocks.{i}.attn.proj.weight", f"data2vec_vision.encoder.layer.{i}.attention.output.dense.weight")
+            (f"{prefix}blocks.{i}.norm1.weight", f"{hf_prefix}encoder.layer.{i}.layernorm_before.weight")
+        )
+        rename_keys.append((f"{prefix}blocks.{i}.norm1.bias", f"{hf_prefix}encoder.layer.{i}.layernorm_before.bias"))
+        rename_keys.append(
+            (f"{prefix}blocks.{i}.attn.proj.weight", f"{hf_prefix}encoder.layer.{i}.attention.output.dense.weight")
         )
         rename_keys.append(
-            (f"{prefix}blocks.{i}.attn.proj.bias", f"data2vec_vision.encoder.layer.{i}.attention.output.dense.bias")
+            (f"{prefix}blocks.{i}.attn.proj.bias", f"{hf_prefix}encoder.layer.{i}.attention.output.dense.bias")
         )
-        rename_keys.append((f"{prefix}blocks.{i}.norm2.weight", f"data2vec_vision.encoder.layer.{i}.layernorm_after.weight"))
-        rename_keys.append((f"{prefix}blocks.{i}.norm2.bias", f"data2vec_vision.encoder.layer.{i}.layernorm_after.bias"))
-        rename_keys.append((f"{prefix}blocks.{i}.mlp.fc1.weight", f"data2vec_vision.encoder.layer.{i}.intermediate.dense.weight"))
-        rename_keys.append((f"{prefix}blocks.{i}.mlp.fc1.bias", f"data2vec_vision.encoder.layer.{i}.intermediate.dense.bias"))
-        rename_keys.append((f"{prefix}blocks.{i}.mlp.fc2.weight", f"data2vec_vision.encoder.layer.{i}.output.dense.weight"))
-        rename_keys.append((f"{prefix}blocks.{i}.mlp.fc2.bias", f"data2vec_vision.encoder.layer.{i}.output.dense.bias"))
+        rename_keys.append(
+            (f"{prefix}blocks.{i}.norm2.weight", f"{hf_prefix}encoder.layer.{i}.layernorm_after.weight")
+        )
+        rename_keys.append((f"{prefix}blocks.{i}.norm2.bias", f"{hf_prefix}encoder.layer.{i}.layernorm_after.bias"))
+        rename_keys.append(
+            (f"{prefix}blocks.{i}.mlp.fc1.weight", f"{hf_prefix}encoder.layer.{i}.intermediate.dense.weight")
+        )
+        rename_keys.append(
+            (f"{prefix}blocks.{i}.mlp.fc1.bias", f"{hf_prefix}encoder.layer.{i}.intermediate.dense.bias")
+        )
+        rename_keys.append((f"{prefix}blocks.{i}.mlp.fc2.weight", f"{hf_prefix}encoder.layer.{i}.output.dense.weight"))
+        rename_keys.append((f"{prefix}blocks.{i}.mlp.fc2.bias", f"{hf_prefix}encoder.layer.{i}.output.dense.bias"))
 
     # projection layer + position embeddings
     rename_keys.extend(
         [
-            (f"{prefix}cls_token", "data2vec_vision.embeddings.cls_token"),
-            (f"{prefix}patch_embed.proj.weight", "data2vec_vision.embeddings.patch_embeddings.projection.weight"),
-            (f"{prefix}patch_embed.proj.bias", "data2vec_vision.embeddings.patch_embeddings.projection.bias"),
+            (f"{prefix}cls_token", f"{hf_prefix}embeddings.cls_token"),
+            (f"{prefix}patch_embed.proj.weight", f"{hf_prefix}embeddings.patch_embeddings.projection.weight"),
+            (f"{prefix}patch_embed.proj.bias", f"{hf_prefix}embeddings.patch_embeddings.projection.bias"),
         ]
     )
 
@@ -74,14 +57,14 @@ def create_rename_keys(config, has_lm_head=False, is_semantic=False):
         # mask token + shared relative position bias + layernorm
         rename_keys.extend(
             [
-                ("mask_token", "data2vec_vision.embeddings.mask_token"),
+                ("mask_token", f"{hf_prefix}embeddings.mask_token"),
                 (
                     "rel_pos_bias.relative_position_bias_table",
-                    "data2vec_vision.encoder.relative_position_bias.relative_position_bias_table",
+                    f"{hf_prefix}encoder.relative_position_bias.relative_position_bias_table",
                 ),
                 (
                     "rel_pos_bias.relative_position_index",
-                    "data2vec_vision.encoder.relative_position_bias.relative_position_index",
+                    f"{hf_prefix}encoder.relative_position_bias.relative_position_index",
                 ),
                 ("norm.weight", "layernorm.weight"),
                 ("norm.bias", "layernorm.bias"),
@@ -101,8 +84,8 @@ def create_rename_keys(config, has_lm_head=False, is_semantic=False):
         # layernorm + classification head
         rename_keys.extend(
             [
-                ("fc_norm.weight", "data2vec_vision.pooler.layernorm.weight"),
-                ("fc_norm.bias", "data2vec_vision.pooler.layernorm.bias"),
+                ("fc_norm.weight", f"{hf_prefix}pooler.layernorm.weight"),
+                ("fc_norm.bias", f"{hf_prefix}pooler.layernorm.bias"),
                 ("head.weight", "classifier.weight"),
                 ("head.bias", "classifier.bias"),
             ]
@@ -111,8 +94,7 @@ def create_rename_keys(config, has_lm_head=False, is_semantic=False):
     return rename_keys
 
 
-# we split up the matrix of each encoder layer into queries, keys and values
-def read_in_q_k_v(state_dict, config, has_lm_head=False, is_semantic=False):
+def read_in_q_k_v(state_dict, config, has_lm_head=False, is_semantic=False, hf_prefix="data2vec_vision."):
     for i in range(config.num_hidden_layers):
         prefix = "backbone." if is_semantic else ""
         # queries, keys and values
@@ -120,25 +102,25 @@ def read_in_q_k_v(state_dict, config, has_lm_head=False, is_semantic=False):
         q_bias = state_dict.pop(f"{prefix}blocks.{i}.attn.q_bias")
         v_bias = state_dict.pop(f"{prefix}blocks.{i}.attn.v_bias")
 
-        state_dict[f"data2vec_vision.encoder.layer.{i}.attention.attention.query.weight"] = in_proj_weight[
+        state_dict[f"{hf_prefix}encoder.layer.{i}.attention.attention.query.weight"] = in_proj_weight[
             : config.hidden_size, :
         ]
-        state_dict[f"data2vec_vision.encoder.layer.{i}.attention.attention.query.bias"] = q_bias
-        state_dict[f"data2vec_vision.encoder.layer.{i}.attention.attention.key.weight"] = in_proj_weight[
+        state_dict[f"{hf_prefix}encoder.layer.{i}.attention.attention.query.bias"] = q_bias
+        state_dict[f"{hf_prefix}encoder.layer.{i}.attention.attention.key.weight"] = in_proj_weight[
             config.hidden_size : config.hidden_size * 2, :
         ]
-        state_dict[f"data2vec_vision.encoder.layer.{i}.attention.attention.value.weight"] = in_proj_weight[
+        state_dict[f"{hf_prefix}encoder.layer.{i}.attention.attention.value.weight"] = in_proj_weight[
             -config.hidden_size :, :
         ]
-        state_dict[f"data2vec_vision.encoder.layer.{i}.attention.attention.value.bias"] = v_bias
+        state_dict[f"{hf_prefix}encoder.layer.{i}.attention.attention.value.bias"] = v_bias
 
         # gamma_1 and gamma_2
         # we call them lambda because otherwise they are renamed when using .from_pretrained
         gamma_1 = state_dict.pop(f"{prefix}blocks.{i}.gamma_1")
         gamma_2 = state_dict.pop(f"{prefix}blocks.{i}.gamma_2")
 
-        state_dict[f"data2vec_vision.encoder.layer.{i}.lambda_1"] = gamma_1
-        state_dict[f"data2vec_vision.encoder.layer.{i}.lambda_2"] = gamma_2
+        state_dict[f"{hf_prefix}encoder.layer.{i}.lambda_1"] = gamma_1
+        state_dict[f"{hf_prefix}encoder.layer.{i}.lambda_2"] = gamma_2
 
         # relative_position bias table + index
         if not has_lm_head:
@@ -147,225 +129,233 @@ def read_in_q_k_v(state_dict, config, has_lm_head=False, is_semantic=False):
             index = state_dict.pop(f"{prefix}blocks.{i}.attn.relative_position_index")
 
             state_dict[
-                f"data2vec_vision.encoder.layer.{i}.attention.attention.relative_position_bias.relative_position_bias_table"
+                f"{hf_prefix}encoder.layer.{i}.attention.attention.relative_position_bias.relative_position_bias_table"
             ] = table
             state_dict[
-                f"data2vec_vision.encoder.layer.{i}.attention.attention.relative_position_bias.relative_position_index"
+                f"{hf_prefix}encoder.layer.{i}.attention.attention.relative_position_bias.relative_position_index"
             ] = index
 
 
-def rename_key(dct, old, new):
-    val = dct.pop(old)
-    dct[new] = val
+def get_args():
+    parser = argparse.ArgumentParser(
+        "Convert Data2VecVision to HF for image classification and pretraining", add_help=False
+    )
+    parser.add_argument("--hf_checkpoint_name", type=str)
+    parser.add_argument("--input_size", default=224, type=int, help="images input size")
+    parser.add_argument("--beit_checkpoint", default="", help="beit checkpoint")
+
+    return parser.parse_args()
 
 
-# We will verify our results on an image of cute cats
-def prepare_img():
-    url = "http://images.cocodataset.org/val2017/000000039769.jpg"
-    im = Image.open(requests.get(url, stream=True).raw)
-    return im
+def load_beit_model(args, is_finetuned, is_large):
+    def load_state_dict(model, state_dict, prefix="", ignore_missing="relative_position_index"):
+        missing_keys = []
+        unexpected_keys = []
+        error_msgs = []
+        # copy state_dict so _load_from_state_dict can modify it
+        metadata = getattr(state_dict, "_metadata", None)
+        state_dict = state_dict.copy()
+        if metadata is not None:
+            state_dict._metadata = metadata
+
+        def load(module, prefix=""):
+            local_metadata = {} if metadata is None else metadata.get(prefix[:-1], {})
+            module._load_from_state_dict(
+                state_dict, prefix, local_metadata, True, missing_keys, unexpected_keys, error_msgs
+            )
+            for name, child in module._modules.items():
+                if child is not None:
+                    load(child, prefix + name + ".")
+
+        load(model, prefix=prefix)
+
+        warn_missing_keys = []
+        ignore_missing_keys = []
+        for key in missing_keys:
+            keep_flag = True
+            for ignore_key in ignore_missing.split("|"):
+                if ignore_key in key:
+                    keep_flag = False
+                    break
+            if keep_flag:
+                warn_missing_keys.append(key)
+            else:
+                ignore_missing_keys.append(key)
+
+        missing_keys = warn_missing_keys
+
+        if len(missing_keys) > 0:
+            print(
+                "Weights of {} not initialized from pretrained model: {}".format(
+                    model.__class__.__name__, missing_keys
+                )
+            )
+        if len(unexpected_keys) > 0:
+            print("Weights from pretrained model not used in {}: {}".format(model.__class__.__name__, unexpected_keys))
+        if len(ignore_missing_keys) > 0:
+            print(
+                "Ignored weights of {} not initialized from pretrained model: {}".format(
+                    model.__class__.__name__, ignore_missing_keys
+                )
+            )
+        if len(error_msgs) > 0:
+            print("\n".join(error_msgs))
+
+    model_kwargs = {
+        "pretrained": False,
+        "use_shared_rel_pos_bias": True,
+        "use_abs_pos_emb": False,
+        "init_values": 0.1,
+    }
+
+    if is_finetuned:
+        model_kwargs.update(
+            {
+                "num_classes": 1000,
+                "use_mean_pooling": True,
+                "init_scale": 0.001,
+                "use_rel_pos_bias": True,
+            }
+        )
+
+    model = create_model(
+        "beit_large_patch16_224" if is_large else "beit_base_patch16_224",
+        **model_kwargs,
+    )
+    patch_size = model.patch_embed.patch_size
+    args.window_size = (args.input_size // patch_size[0], args.input_size // patch_size[1])
+    checkpoint = torch.load(args.beit_checkpoint, map_location="cpu")
+
+    print(f"Load ckpt from {args.beit_checkpoint}")
+    checkpoint_model = None
+    for model_key in ("model", "module"):
+        if model_key in checkpoint:
+            checkpoint_model = checkpoint[model_key]
+            print(f"Load state_dict by model_key = {model_key}")
+            break
+
+    all_keys = list(checkpoint_model.keys())
+    for key in all_keys:
+        if "relative_position_index" in key:
+            checkpoint_model.pop(key)
+
+        if "relative_position_bias_table" in key:
+            rel_pos_bias = checkpoint_model[key]
+            src_num_pos, num_attn_heads = rel_pos_bias.size()
+            dst_num_pos, _ = model.state_dict()[key].size()
+            dst_patch_shape = model.patch_embed.patch_shape
+            if dst_patch_shape[0] != dst_patch_shape[1]:
+                raise NotImplementedError()
+
+    load_state_dict(model, checkpoint_model, prefix="")
+
+    return model
 
 
-@torch.no_grad()
-def convert_data2vec_vision_checkpoint(checkpoint_url, pytorch_dump_folder_path):
-    """
-    Copy/paste/tweak model's weights to our Data2VecVision structure.
-    """
+def main():
+    args = get_args()
 
-    # define default Data2VecVision configuration
+    is_finetuned = "ft1k" in args.hf_checkpoint_name
+    is_large = "large" in args.hf_checkpoint_name
+
+    if is_finetuned:
+        # To convert Beit's data2vec_vision to HF you need to copy
+        # https://github.com/facebookresearch/data2vec_vision/blob/main/beit/modeling_finetune.py
+        # into this folder.
+        import modeling_finetune  # noqa: F401
+    else:
+        # To convert Beit's data2vec_vision to HF you need to copy
+        # https://github.com/facebookresearch/data2vec_vision/blob/main/beit/modeling_cyclical.py
+        # into this folder
+        # IMPORTANT: Note that for now we've only converted the down-stream
+        # model and not the full pretrained model. This means for the integration
+        # test you need to add a `return x` after the following line:
+        # https://github.com/facebookresearch/data2vec_vision/blob/af9a36349aaed59ae66e69b5dabeef2d62fdc5da/beit/modeling_cyclical.py#L197
+        # to make the integration test pass.
+        import modeling_cyclical  # noqa: F401
+
+    # 1. Create model config
     config = Data2VecVisionConfig()
-    has_lm_head = False
-    is_semantic = False
-    repo_id = "datasets/huggingface/label-files"
-    # set config parameters based on URL
-    if checkpoint_url[-9:-4] == "pt22k":
-        # masked image modeling
-        config.use_shared_relative_position_bias = True
-        config.use_mask_token = True
-        has_lm_head = True
-    elif checkpoint_url[-9:-4] == "ft22k":
-        # intermediate fine-tuning on ImageNet-22k
+    if is_finetuned:
         config.use_relative_position_bias = True
-        config.num_labels = 21841
-        filename = "imagenet-22k-id2label.json"
-        id2label = json.load(open(hf_hub_download(repo_id, filename), "r"))
-        id2label = {int(k): v for k, v in id2label.items()}
-        # this dataset contains 21843 labels but the model only has 21841
-        # we delete the classes as mentioned in https://github.com/google-research/big_transfer/issues/18
-        del id2label[9205]
-        del id2label[15027]
-        config.id2label = id2label
-        config.label2id = {v: k for k, v in id2label.items()}
-    elif checkpoint_url[-8:-4] == "to1k":
-        # fine-tuning on ImageNet-1k
-        config.use_relative_position_bias = True
+        config.use_shared_relative_position_bias = False
+        config.use_mean_pooling = True
         config.num_labels = 1000
+
+        repo_id = "datasets/huggingface/label-files"
         filename = "imagenet-1k-id2label.json"
         id2label = json.load(open(hf_hub_download(repo_id, filename), "r"))
         id2label = {int(k): v for k, v in id2label.items()}
         config.id2label = id2label
         config.label2id = {v: k for k, v in id2label.items()}
-        if "384" in checkpoint_url:
-            config.image_size = 384
-        if "512" in checkpoint_url:
-            config.image_size = 512
-    elif "ade20k" in checkpoint_url:
-        # fine-tuning
-        config.use_relative_position_bias = True
-        config.num_labels = 150
-        filename = "ade20k-id2label.json"
-        id2label = json.load(open(hf_hub_download(repo_id, filename), "r"))
-        id2label = {int(k): v for k, v in id2label.items()}
-        config.id2label = id2label
-        config.label2id = {v: k for k, v in id2label.items()}
-        config.image_size = 640
-        is_semantic = True
     else:
-        pass
-#        raise ValueError("Checkpoint not supported, URL should either end with 'pt22k', 'ft22k', 'to1k' or 'ade20k'")
+        config.use_relative_position_bias = False
+        config.use_shared_relative_position_bias = True
+        config.use_mean_pooling = False
 
-    # size of the architecture
-    if "base" in checkpoint_url:
-        pass
-#    elif "large" in checkpoint_url:
-#        config.hidden_size = 1024
-#        config.intermediate_size = 4096
-#        config.num_hidden_layers = 24
-#        config.num_attention_heads = 16
-#        if "ade20k" in checkpoint_url:
-#            config.image_size = 640
-#            config.out_indices = [7, 11, 15, 23]
-#    else:
-#        raise ValueError("Should either find 'base' or 'large' in checkpoint URL")
+    if is_large:
+        config.hidden_size = 1024
+        config.intermediate_size = 4096
+        config.num_hidden_layers = 24
+        config.num_attention_heads = 16
 
-    # load state_dict of original model, remove and rename some keys
-    state_dict = torch.hub.load_state_dict_from_url(checkpoint_url, map_location="cpu", check_hash=True)
-    state_dict = state_dict["model"] if "ade20k" not in checkpoint_url else state_dict["state_dict"]
+    # 2. Load Beit model
+    orig_model = load_beit_model(args, is_finetuned, is_large)
+    orig_model.eval()
 
-    rename_keys = create_rename_keys(config, has_lm_head=has_lm_head, is_semantic=is_semantic)
-    for src, dest in rename_keys:
-        rename_key(state_dict, src, dest)
-    read_in_q_k_v(state_dict, config, has_lm_head=has_lm_head, is_semantic=is_semantic)
-    if is_semantic:
-        # add prefix to decoder keys
-        for key, val in state_dict.copy().items():
-            val = state_dict.pop(key)
-            if key.startswith("backbone.fpn"):
-                key = key.replace("backbone.fpn", "fpn")
-            state_dict[key] = val
-
-    # load HuggingFace model
-    if checkpoint_url[-9:-4] == "pt22k":
-        model = Data2VecVisionForMaskedImageModeling(config)
-    elif "ade20k" in checkpoint_url:
-        model = Data2VecVisionForSemanticSegmentation(config)
-    else:
-        model = Data2VecVisionForImageClassification(config)
-    model.eval()
-    model.load_state_dict(state_dict)
-
-    # Check outputs on an image
-    if is_semantic:
-        feature_extractor = BeitFeatureExtractor(size=config.image_size, do_center_crop=False)
-        ds = load_dataset("hf-internal-testing/fixtures_ade20k", split="test")
-        image = Image.open(ds[0]["file"])
-    else:
-        feature_extractor = BeitFeatureExtractor(size=config.image_size, resample=Image.BILINEAR, do_center_crop=False)
-        image = prepare_img()
-
+    # 3. Forward Beit model
+    feature_extractor = BeitFeatureExtractor(size=config.image_size, do_center_crop=False)
+    image = Image.open("../../../../tests/fixtures/tests_samples/COCO/000000039769.png")
     encoding = feature_extractor(images=image, return_tensors="pt")
     pixel_values = encoding["pixel_values"]
 
-    outputs = model(pixel_values)
-    logits = outputs.logits
+    orig_args = (pixel_values,) if is_finetuned else (pixel_values, None)
+    with torch.no_grad():
+        orig_model_output = orig_model(*orig_args)
 
-    # verify logits
-    expected_shape = torch.Size([1, 1000])
-    if checkpoint_url[:-4].endswith("data2vec_vision_base_patch16_224_pt22k"):
-        expected_shape = torch.Size([1, 196, 8192])
-    elif checkpoint_url[:-4].endswith("data2vec_vision_large_patch16_224_pt22k"):
-        expected_shape = torch.Size([1, 196, 8192])
-    elif checkpoint_url[:-4].endswith("data2vec_vision_base_patch16_224_pt22k_ft22k"):
-        expected_shape = torch.Size([1, 21841])
-        expected_logits = torch.tensor([2.2288, 2.4671, 0.7395])
-        expected_class_idx = 2397
-    elif checkpoint_url[:-4].endswith("data2vec_vision_large_patch16_224_pt22k_ft22k"):
-        expected_shape = torch.Size([1, 21841])
-        expected_logits = torch.tensor([1.6881, -0.2787, 0.5901])
-        expected_class_idx = 2396
-    elif checkpoint_url[:-4].endswith("data2vec_vision_base_patch16_224_pt22k_ft1k"):
-        expected_logits = torch.tensor([0.1241, 0.0798, -0.6569])
-        expected_class_idx = 285
-    elif checkpoint_url[:-4].endswith("data2vec_vision_base_patch16_224_pt22k_ft22kto1k"):
-        expected_logits = torch.tensor([-1.2385, -1.0987, -1.0108])
-        expected_class_idx = 281
-    elif checkpoint_url[:-4].endswith("data2vec_vision_base_patch16_384_pt22k_ft22kto1k"):
-        expected_logits = torch.tensor([-1.5303, -0.9484, -0.3147])
-        expected_class_idx = 761
-    elif checkpoint_url[:-4].endswith("data2vec_vision_large_patch16_224_pt22k_ft1k"):
-        expected_logits = torch.tensor([0.4610, -0.0928, 0.2086])
-        expected_class_idx = 761
-    elif checkpoint_url[:-4].endswith("data2vec_vision_large_patch16_224_pt22k_ft22kto1k"):
-        expected_logits = torch.tensor([-0.4804, 0.6257, -0.1837])
-        expected_class_idx = 761
-    elif checkpoint_url[:-4].endswith("data2vec_vision_large_patch16_384_pt22k_ft22kto1k"):
-        expected_logits = torch.tensor([[-0.5122, 0.5117, -0.2113]])
-        expected_class_idx = 761
-    elif checkpoint_url[:-4].endswith("data2vec_vision_large_patch16_512_pt22k_ft22kto1k"):
-        expected_logits = torch.tensor([-0.3062, 0.7261, 0.4852])
-        expected_class_idx = 761
-    elif checkpoint_url[:-4].endswith("data2vec_vision_base_patch16_640_pt22k_ft22ktoade20k"):
-        expected_shape = (1, 150, 160, 160)
-        expected_logits = torch.tensor(
-            [
-                [[-4.9225, -2.3954, -3.0522], [-2.8822, -1.0046, -1.7561], [-2.9549, -1.3228, -2.1347]],
-                [[-5.8168, -3.4129, -4.0778], [-3.8651, -2.2214, -3.0277], [-3.8356, -2.4643, -3.3535]],
-                [[-0.0078, 3.9952, 4.0754], [2.9856, 4.6944, 5.0035], [3.2413, 4.7813, 4.9969]],
-            ]
-        )
-    elif checkpoint_url[:-4].endswith("data2vec_vision_large_patch16_640_pt22k_ft22ktoade20k"):
-        expected_shape = (1, 150, 160, 160)
-        expected_logits = torch.tensor(
-            [
-                [[-4.3305, -2.3049, -3.0161], [-2.9591, -1.5305, -2.2251], [-3.4198, -1.8004, -2.9062]],
-                [[-5.8922, -3.7435, -4.3978], [-4.2063, -2.7872, -3.4755], [-4.2791, -3.1874, -4.1681]],
-                [[0.9895, 4.3467, 4.7663], [4.2476, 5.6830, 6.1518], [4.5550, 6.2495, 6.5154]],
-            ]
-        )
+    # 4. Load HF Data2VecVision model
+    if is_finetuned:
+        hf_model = Data2VecVisionForImageClassification(config)
+        hf_model.eval()
+        has_lm_head = False
+        hf_prefix = "data2vec_vision."
     else:
-        raise ValueError("Can't verify logits as model is not supported")
+        hf_model = Data2VecVisionModel(config)
+        hf_model.eval()
+        has_lm_head = True
+        hf_prefix = ""
 
-    assert logits.shape == expected_shape, "Shape of logits not as expected"
-    if not has_lm_head:
-        if is_semantic:
-            assert torch.allclose(
-                logits[0, :3, :3, :3], expected_logits, atol=1e-3
-            ), "First elements of logits not as expected"
-        else:
-            print("Predicted class idx:", logits.argmax(-1).item())
-            assert torch.allclose(
-                logits[0, :3], expected_logits, atol=1e-3
-            ), "First elements of logits not as expected"
-            assert logits.argmax(-1).item() == expected_class_idx, "Predicted class index not as expected"
+    rename_keys = create_rename_keys(config, hf_prefix=hf_prefix, has_lm_head=has_lm_head)
+    state_dict = orig_model.state_dict()
+    for src, dest in rename_keys:
+        val = state_dict.pop(src)
+        state_dict[dest] = val
 
-    Path(pytorch_dump_folder_path).mkdir(exist_ok=True)
-    print(f"Saving model to {pytorch_dump_folder_path}")
-    model.save_pretrained(pytorch_dump_folder_path)
-    print(f"Saving feature extractor to {pytorch_dump_folder_path}")
-    feature_extractor.save_pretrained(pytorch_dump_folder_path)
+    read_in_q_k_v(state_dict, config, hf_prefix=hf_prefix, has_lm_head=has_lm_head)
+    missing_keys, unexpected_keys = hf_model.load_state_dict(state_dict, strict=False)
+    print("HF missing", missing_keys)
+    print("HF unexpected_keys", unexpected_keys)
+
+    # 5. Forward HF Data2VecVision model
+    with torch.no_grad():
+        hf_model_output = hf_model(pixel_values)
+
+    hf_output = hf_model_output.logits if is_finetuned else hf_model_output.last_hidden_state
+
+    # 6. Compare
+    max_absolute_diff = torch.max(torch.abs(hf_output - orig_model_output)).item()
+
+    print(f"max_absolute_diff = {max_absolute_diff}")
+    success = torch.allclose(hf_output, orig_model_output, atol=1e-3)
+    print("Do both models output the same tensors?", "🔥" if success else "💩")
+    if not success:
+        raise Exception("Something went wRoNg")
+
+    # 7. Save
+    print(f"Saving to {args.hf_checkpoint_name}")
+    hf_model.save_pretrained(args.hf_checkpoint_name)
+    feature_extractor.save_pretrained(args.hf_checkpoint_name)
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-
-    parser.add_argument(
-        "--checkpoint_url",
-        default="https://unilm.blob.core.windows.net/data2vec_vision/data2vec_vision_base_patch16_224_pt22k_ft22kto1k.pth",
-        type=str,
-        help="URL to the original PyTorch checkpoint (.pth file).",
-    )
-    parser.add_argument(
-        "--pytorch_dump_folder_path", default=None, type=str, help="Path to the folder to output PyTorch model."
-    )
-    args = parser.parse_args()
-    convert_data2vec_vision_checkpoint(args.checkpoint_url, args.pytorch_dump_folder_path)
+    main()
